+++ src/asset.rs
@@ -1,26 +1,75 @@
+use regex::Regex;
 use std::ffi::OsStr;
 use std::fs;
 use std::fs::DirEntry;
 use std::fs::File;
 use std::io::BufRead;
 use std::io::BufReader;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 #[derive(Clone)]
 pub struct Asset {
     pub extension: Option<String>,
-    pub hash: String,
-    pub path_name: String,
+    pub guid: String,
+    pub path: String,
     pub has_meta: bool,
+    pub asset_type: AssetType,
+}
+
+#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone)]
+pub enum AssetType {
+    FbxModel,
+    Material,
+    Prefab,
+    Scene,
+    Other(String),
 }
 
 impl Asset {
-    pub fn from_path(entry: &DirEntry) -> Option<Asset> {
+    pub fn try_get_mat_texture_guid(&self) -> Option<String> {
+        match &self.asset_type {
+            AssetType::Material => {}
+            _ => return None,
+        }
+        let file = File::open(&self.path).unwrap();
+        let buf_reader = BufReader::new(file);
+        let search = buf_reader.lines().find(|s| {
+            let ss = s.as_ref().unwrap();
+            ss.contains("m_Texture") && ss.contains("guid: ")
+        });
+        if let Some(line) = search {
+            let line = line.unwrap_or_default();
+            return extract_guid(&line);
+        }
+        None
+    }
+
+    pub fn prepare_directory(&self) {
+        println!("{}: {:?}", self.guid, self.path);
+        let base_path = Path::new(&self.path);
+        let result_dir = base_path.parent();
+        if result_dir.is_none() {
+            eprintln!("{} is none", &self.path);
+        }
+        let result_dir = result_dir.unwrap();
+        if !result_dir.exists() {
+            let result = fs::create_dir_all(result_dir);
+            if result.is_err() {
+                eprintln!(
+                    "Error {}: {}",
+                    result_dir.to_str().unwrap(),
+                    result.err().unwrap()
+                );
+            }
+        }
+    }
+
+    pub fn from_path(entry: &DirEntry, output_dir: &Path) -> Option<Asset> {
         let root_file = entry.path();
         if !root_file.is_dir() {
             return None;
         }
-        let asset = entry.file_name().into_string().unwrap();
+        let guid = entry.file_name().into_string().unwrap();
         let mut real_path = String::new();
         let mut extension = None;
         let mut has_asset = false;
@@ -36,7 +85,7 @@ impl Asset {
                     let line = buf_reader.lines().next();
                     match line {
                         Some(Ok(path)) => {
-                            real_path = path;
+                            real_path = output_dir.join(path).to_str().unwrap().to_string();
                             if let Some(e) =
                                 Path::new(&real_path).extension().and_then(OsStr::to_str)
                             {
@@ -52,14 +101,31 @@ impl Asset {
             }
         }
         if has_asset {
+            let asset_type = match &extension {
+                Some(str) => match str.as_str() {
+                    "fbx" => AssetType::FbxModel,
+                    "prefab" => AssetType::Prefab,
+                    "unity" => AssetType::Scene,
+                    "mat" => AssetType::Material,
+                    _ => AssetType::Other(str.clone()),
+                },
+                _ => AssetType::Other(String::new()),
+            };
             Some(Asset {
                 extension,
-                hash: asset,
-                path_name: real_path,
+                guid,
+                path: real_path,
                 has_meta,
+                asset_type,
             })
         } else {
             None
         }
     }
+}
+
+fn extract_guid(text: &str) -> Option<String> {
+    let re = Regex::new(r"guid: (?P<guid>[A-Za-z0-9]{32})").unwrap();
+    re.captures(text)
+        .and_then(|cap| cap.name("guid").map(|guid| guid.as_str().to_string()))
 }